Updated README Deploying with Docker in Windows (#2598)
Updated instructutions for running Beacon Node in Docker on Windows. The process requires a few additional steps to make a local volume available to mount the data dir. If this doesn't happen, the /tmp/prysm-data dir won't be created at run, and the chaindata will not be stored locally, nor will the account creation process store the key files appropriately for reference when running the validator.