Implement `SaveLowestSignedSourceEpochForPubKey`
This function is useful when, working with minimal anti-slashing DB,
we want to override the lowest signed source epoch, in case the
interchange file contains itself some surrounding attestations.

Example:
- Attestation 1: S1 =========> T1
- Attestation 2:         S2 => T2

In this case, we won't accept attestation 2, but we still want to
save S2 as lowest signed source epoch, so we won't accept
any incoming attestation with source < S2.