Only start from the skipped slot cache if better
ProcessSlots takes a state from the skipped slot cache and assumes that it's better than the passed current state.

If the incoming state.Slot() is higher than the skipped slot cache, then the highest slot is set to be the cache, we process slots again for the same state and then we put back to the cache.